+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 #### 0.3.12 - 2016-XX-XX
-* Optimized `MVar.read`.
+* Optimized `MVar.read`, `MVar.mutateFun` and `MVar.modifyFun`.
 
 #### 0.3.11 - 2016-08-20
 * Added `MVar.mutateFun`, `MVar.mutateJob`, `MVar.tryMutateFun`,